This comments out some lines in trainer.py and model.py so that training and execution all function while still outputing the JSON tensor file. However some files are not output in this version.
The word encoding and the render process can be seperated now, there are different functions for each. This is accessed by setting makeWords and renderIt to true, which can also be done on command line. The run.sh is updated to do this.

Be aware that there are several tensors output to JSON, and not all are read back, please check if the tensor you want to operate on is coming from a file or another source by looking in model.py or trainer.py.

written:
out_working_feature
out_style_feature

read:
B_working_feature
in_working_feature
in_style_feature

Interpolation of tensors has been moved to a new function. There is also extranious code to read a 'default' or neutral feature tensor, for the birds dataset this is called plain_bird. As well there is a smile_bird tensor file.